Opening SV on MDHisto with attached MDE acts as if MDE was clicked
It was requested that if the MDE data is available then an MDHistoWorkspace can
show bin controls to rebin the data. The simplest way of doing this is to
have the sliceviewer swap the given workspace for the original MDE and
work off this. The full rebinning works as expected then. The window title
is updated to reflect this and to notify the user what has happened.