Abins: limit reader check of advanced_parameters
The test of advanced_parameters is needlessly sensitive to instrument
parameters that do not impact the test in question. Allow specific
instruments of interest to be specified.

This code will be removed in the next commit for reasons given
there. Committing it to the history in case it becomes relevant in
future.