Abins: remove advance_parameters check from loader tests
The unit tests for Abins include a standardised check that data from
many ab initio data files is imported correctly. The resulting object
is compared to data from reference .dat files.

These files also contain a cache of the "advanced_parameters" at the
point they were generated. This results in a need to regenerate test
data whenever the advanced parameter defaults are
changed/renamed/restructured.

Presumably the original purpose of this check was to ensure that
parameters used in the file loading process are unchanged. However, at
present abins.parameters is not used at all by the file loaders! There
is apparently no value in validating these parameters when the purpose
of the test is to ensure that the file parsing and data object
instantiation are working.

This commit therefore removes that check, and the code for writing
such data to reference files. Old reference files will continue to
pass tests as the presence of extra dict keys is tolerated.