Use RefRoi for conversion to Q
This commit replaces the current mixture of methods for conversion to Q with a single method using the child algorithm RefRoi. This fixes the workflow algorithm history when summing in lambda because previously this calculation was done within this algorithm rather than via a child algorithm. It also simplifies the code in that the same method can be used for both summation methods.

This commit also adds validation that ThetaIn is provided when summing in lambda. This is always the case when running this algorithm via ReflectometryReductionOneAuto.

Re 23012