Abins 2D as a separate Algorithm
This is an initial implementation of a separate "Abins2D" interface,
based on the version built for a demonstration to ISIS Molecular
Spectroscopy scientists.

- Further user testing is needed to determine the appropriate feature
  set. Suggestions already under consideration include:

  - Instrument setting/chopper options
  - Incident energy selection

The free entry of resolution value should be removed and replaced with
a reasonable selection of instruments/setups with appropriate set
values. A natural first target will be MAPS.

Initial testing shows a curious behaviour in the current
implementation, which reduces overall intensity as the resolution
width is increased.