Generate doctrings lazily on request
Alongside generating the signature when it is reqested, we also have to
wait to generate the docstring for the algorithm since the list of
properties is required and the algorithm has to be initialised to get
them.
This currently doesn't help so much with the Workbench, as it creates a
full list of the signatures and docstrings for the code editor
completion. But will help with import times of the simpleapi when using
mantid python, for example.

RE #28809