Improve and expand unit tests for angle corrections
Add additional test cases to cover all combinations of the type of summation and whether ThetaIn is specified or not (i.e. whether angle correction is requested or not).
Tidy the code and add supporting functions to make it clearer which angle is being tested against (i.e. the provided angle or the detector angle).

Re #23012