Refactored re-plotting
There are two modes in which plotting can operate in the Muon Gui's. One mode shows the groups/pairs derived from the raw data and one shows the workspaces which are currently being fitted. This is controlled by which tab you are on. In the case where you have undocked the tabs such as this concept doesn't make sense a combo box has been added so you can manually switch between the two modes.

A further change is that the ability to select plot type has been disabled from plotting widget when you are in fitting mode as in this case you want the plots to always show the data currently being fitted and being able to control what was plotted from two different locations led to confusion.