Abins: rework SData to take frequencies as a separate argument
This object can be worked with a bit more reasonably if the atom data
is separated from the frequency list that applies to all of them.

We can still package them in a dict for caching (for now) but it would
make more sense to treat the atom data as list, which can be the next
step in this rework.