Improve the error calculation done in MCAbsorptionStrategy
The error on each point in the attenuation factor spectrum was previously
calculated as 1/sqrt(nevents). This error was the same for all points
This has been changed to the standard deviation/sqrt(nevents) which is the
appropriate calculation given that each point in the spectrum is a mean of the
attenuation factors calculated over various simulated tracks through the sample

At this stage, the error is only output if you're not using a sparse instrument.
The spatial interpolation used in the sparse calculation means the errors need
to be propagated through the interpolation calculation which needs further work