Abins SData: remove bin_width input argument (long commit message)
SData is the Abins data container for atom-resolved scattering
intensity data. Pre-refactor, it essentially wraps a dict of the form

{'frequencies': np.ndarray
 'atom_1': {'s': {'order_1': np.ndarray, 'order_2': np.ndarray, ...}}}

There are a couple of curiosities here:
- It's not clear what per-atom data would be included other than
  's'. The class is named SData. So this dict level seems redundant.
- Looking up atoms with strings is rather awkward compared to using
  list indices. The only benefit of using a dict is that 'frequencies'
  can be packaged in the same object. While that's handy for
  serialization, it doesn't make a lot of sense to treat it this way
  in a special class.

Packaging the data for different quantum orders as dict elements is
quite nice as it means one can safely and efficiently check for
existence, access them randomly and update/append in place.

In addition there are two bits of metadata attached to this class:
bin_width and temperature.

Bin width is handled purely for sanity-checking and is not actually an
essential part of the data. While it is true that Abins currently
works with regular energy bins, it is not clear that this should be
enforced at the data class level; irregularly-spaced S data is still S
data and this class is otherwise equipped to handle such data, so why
prevent it from being used in such scenarios?

We therefore remove this mandatory constraint, and instead define a
getter for bin_width that derives a bin_width from the frequency data,
returning None if the data is not evenly spaced. This method can be
called from a higher level if the even-spacing constraint is to be
enforced.

The "sample_form" metadata should be considered an arbitrary string
from the perspective of this data container. Synthetic data might not
have a reference physical state, so again we should allow an empty entry.

The remaining piece of metadata is the temperature. This is relevant
information for cache invalidation or even import/plotting as the S
spectrum should be associated with a temperature. If S was calculated
in a model neglecting temperature effects, this might be denoted with
a temperature of zero or None. Again, it makes more sense for this to
be enforced a higher level, with the data class simply polled for a
consistency check. So we remove the mandatory check/error for valid
temperature and replace it with a more polite method that may be
called as needed.

As it happens, the sanity checks are mostly _not_ needed by
spowdersemiempirical calculator as it generates these inputs in
acceptable form (or inherits them from other parts of Abins). But we
can do a bin-width check on the data when loaded from cache. This
means spowdersemiempiricalcalculator no longer needs to set bin_width
as a private property; it is no longer used outside its __init__().