Abins SData: support index access
Allow the cumbersome extract() dict to be bypassed and directly index
items from an Abins SData object. Slicing is also supported. Test the
new functionality before using it elsewhere.

This will enable cleaner interactions with this object in the main
Abins algorithm. Eventually the extract() method should be repurposed
such that it is only used for serialisation.