Abins kpointsdata: unpack input data to separate input arguments
Now that these data arrays are the only input to instantiate the
class, it doesn't make sense to require them packed into a dict. If
one has a dict already, they can use the **kwargs syntax.

Something funny has been going on with the unit_cell input; this was
not marked as optional anywhere, but some of the tests were able to
neglect it. This would have been raising the ValueError that the tests
were looking for for, skipping the intended test operation. This has
been rectified!

It's not really clear why it should exist here at all; we
still need the atomic positions to do most structure-related things,
which is stored in AtomsData. Why not keep unit_cell there?

The tests are also refactored to cut redundancy and make them easier
to read.