Abins kpointsdata: remove instantiation from dict
The from_dict instantiation is not actually useful any more because
the required dictionary format (numpy arrays) does not match the
output of extract(). We don't need to write checks that all the
appropriate keys are present because Python's mandatory named
arguments syntax [ __init__(self, *, a=..., b=..., ...)  ] is
enforcing this now.

A good from_dict would be useful, but we'll get there when the rest of
the API is better.