Abins KpointsData: index with int rather than string
KPointsData stores data by k-point in dicts indexed by str(i) for i in
range(nkpts). This is not the most obvious choice of data structure.

The "obvious" option would be numpy arrays with an axis over
kpts (i.e. the format used for init!) Presumably this option was
dropped because KpointsData can store a different number of
frequencies at each k-point, whereas numpy arrays are not generally
"ragged". It might be possible to implement it by pairing a padded
array with a mask; at least we know _a priori_ that rows will not
exceed 3N frequencies.

A natural way to achieve ragged data storage while getting good
performance within each loop iteration over a k-point is to use a
Python list of 1-D arrays.

Instead of that, the current implementation uses a dict. Compared to
lists, Python dicts provide faster random access at the cost of
increased memory usage. So this may have been done for performance
reasons? It is not clear however that Abins will ever need to perform
intensive random access over k-points; iteration is the main operation
that matters and here there should be less of a performance
difference.

A further detail is that the dict keys are string representations of
integer indices. This leads to a lot of string casting and explicit
enumeration in the code that works with this object; it's not very
pretty. The main use case of string keys would be JSON serialisation
but a) that isn't implemented (yet) and b) it could be handled
at load/dump time with a custom parser. (Loading/dumping numpy arrays
requires some kind of conversion step to be in place anyway.)

So as a precursor to changing the data storage to an array or list
system, change the dict indexing to bare integers and fix whatever
that breaks. This makes some of the code that works with the extracted
dict slightly uglier at this point, as we redefine the
abins.constants.GAMMA_POINT to an int. In general we aim to discourage
use of the extracted dictionaries and favour the object API.