Add executable launcher for workbench on windows/macOS
The startup scripts for macOS/Windows are nearly indentical.
Harmonising them lowers the maintenance burden. To be portable
a simple executable has been created that performs the tasks
of the old launch scripts.

On Windows this has the advantage of removing the .exe that
was 'compiled' by powershell and flagged as a virus by some
scanners such as McAfee. It also removes any requirement for
the .NET framework to be available.

The error reporter code has been moved to mantidqt and gained
a main module that can be invoked with -m, thereby greatly
simplifying finding and starting the dialog.