Load python completion tooltips on request
Changes the loading of the code completions to only start loading the
ones from the simpleapi once some changes happen in the code editor.
This will reduce startup times, especially if you don't need to use the
code editor. Also loads the completions in a thread, and releases the
GIL for algorithm initialization, so it does not freeze the GUI.

RE #29002