Abins: acoustic threshold from kpointsdata to powdercalculator
The semi-analytic model does not handle imaginary modes well, so they
are removed before calculating intensities. However, the way this is
currently managed is disorientating:

- The atomic displacements and frequencies stored in kpointsdata are
  trimmed to remove modes below ACOUSTIC_PHONON_THRESHOLD
- But the full set of calculated frequencies is provided to
  SPowderSemiEmpiricalCalculator, and must also be trimmed in order
  to match dimensions with the results of PowderCalculator (which
  computes the tensors needed for the semi-analytic averaging method.)

The deficiencies of this approach are

- Discarding raw data prematurely, when other analysis methods and
  visualisation choices might want to see or renormalise those imaginary
  modes.
- Redundancy, by thresholding in more than one part of the code.

In order to address this, we move the responsibility for thresholding
into PowderCalculator/PowderData, allowing the tensor calculation to
apply a threshold and return a commensurate set of frequencies
alongside the tensor data. SPowderSemiEmpiricalCalculator can then
work directly with these frequencies, and need not deal with the raw
frequency set.

It is unclear why the acoustic threshold is in abins.constants rather
than abins.parameters, given that its value is somewhat
arbitrary. Perhaps this should be addressed within a general review of
the abins.constants file.

The extract() method of powderdata is modified so that while integer
keys are preferred for the k-points, they are cast to str on
extract(). This is done to support the current caching system,
which uses these keys inside an hdf5 path. While it would be possible
to implement a more complete path to/from serialisable data, we prefer
to defer this to a separate package of work while considering that
system as a whole.

This commit breaks the data loading tests, presumably because the
reference data has clipped off the zero/imaginary modes. Following
commits will test this assumption before replacing the reference data.