Update view and fields on user input
On the mask tab of the instrument view, when the user inputs a value in
one the field, the view is now updated on validation of this value (i.e.
when the user clicks outside the field or hits Enter), whereas before,
the user needed to hover over the instrument view to see the changes
take effect, which was illogical.

Also, when there is a correction on the value entered (for example a
negative value for a positive-only field), the value corrected is
immediatly showed, instead of waiting for the user to deselect-reselect
the shape.