Forbid the direct modifying of the bounding box
The user no longer edit the bounding box fields for the sector mask,
because this is not a well-defined operation (there is a lot of
different new sectors that correspond to a given field edit, and it's
not possible to decide which one the user meant (nor it is possible for
the user to mean one at all, in fact)).

Thus the user can enter new value on the fields, but upon leaving them /
validating, they will be reversed to their previous values.