Abins KpointsData: add iteration over KpointData namedtuple
Abins generally works on one k-point at a time, so a natural interface
for the KpointsData is to iterate, yielding data for each
k-point. This is encapsulated in a NamedTuple to provide a simple
readable interface without too much code.

len(kpoints_data) then becomes a very natural way of getting the
number of k-points.

This interface is then used in the PowderCalculator, eliminating a
call to KpointsData.extract() and iterating over the object in a more
Pythonic way.