Abins: spowdersemiempiricalcalculator working directly with objects
Remove any dependence of SPowderSemiEmpiricalCalculator on extract()
methods except when caching its own results.

As powder data is currently drawn from cache file rather than passed
as an argument, we provide a classmethod on PowderData so that this
dict can immediately be converted to a full PowderData object and its
API used appropriately.

This commit also switches to reading the fundamental frequency set
from PowderData, eliminating an independent comparison to
ACOUSTIC_PHONON_THRESHOLD.

It looks like one goal of the caching/extract() setup was to allow
Abins to work with data from only one k-point at a time to save
memory. However, the implementation here doesn't seem very efficient;
on each iteration, data is loaded for all k-points and then the
unwanted data is discarded. It might be worth taking this into account
if/when reworking the cache by providing a way to inexpensively load
data for a single k-point at the HDF5 level. This could involve
changing the stored data heirarchy.