Correct return type in sip declaration
Setting up a new macOS builder gave segfaults
when MantidPlot was shutting down. The debugger
pointed to the interpreter dock already having
been deleted by Python before the C++ closeEvent
line 'delete m_interpreterDock'. The assumption
is that the return type not being marked as a
pointer confused sip and forced it to take
ownership and delete it. We fix the return
type and this has fixed the segfault