Remove embedded Python.app Info.plist files
We want the mac bundle to just appear to be workbench
but as we start Python to do this the application name
on mac (next to the Apple logo) displays Python if
these files are left in. Starting Python standalone from
the terminal ignores these files anyway so they are of
no use to us.