Improve unit tests to use larger test workspace
In readiness for changes to summing in Q we will need unit tests using a workspace with 5 detectors instead of 3 so we can sum a region that has adjacent detectors. This commit updates the workspaces used in the unit tests to add an extra couple of spectra and adjusts values in them such that the middle 3 detectors are at the same position with the same values as the original 3 spectra, so that the results stay the same.

Because there is one additional detector before the main detectors of interest, the processing instructions have been incremented by 1 for the tests. Note that some of the test names had incorrect numbers so they look like they have been incremented by 2, but they are just made consistent with what the test is actually doing.

Re #23012