Removed uneeded signal Re #29015
The results tab was being informed that a fit was finishing via two different routes. First the model was updating the selected fit function and then the presenter was updating the selected fit function and the displayed fits. These two signals were not both uneecary and were interfering with each other so have removed the signal in the model