Move directory iteration into getIPFFilename method
Moves iterating through all directories where an IPF could be into the
method which looks this up. This means a caller can still pass a
directory hint, which we will look at first.

But afterwards we will iterate over instrument directories on their
behalf, avoiding them needing to do this