Introduce package SITE_PACKAGES CMake variable
Allows the location of the installed Python packages
to be updated more easily in the future. This is not
so easy while currently bundling MantidPlot/MantidWorkbench
together but once MantidPlot disappears it will allow
modification of the packaged location more easily.

Also introduces a CMake variable to define the tag
used by the dynamic loader to identify the path of the
loading resource. Reduces copy and paste of setting
RPATH entities in the Python framework