Add grouping option for pixel and tube
Adds the possibility to select and group individual pixel and tube.
HOWEVER due to these selections being dependent of the detector, contrary
to every other shape used for masking etc, the selection cannot be
printed on screen, as it would require to create a shape that follows
the detector if it is moved and is rendered using the correct
perspective.
Moreover, contrary to what happens with masking and ROIing, there are
two options for grouping, which means the operation cannot be applied
immediatly on click in this case.
Since this means that a) the user should input after doing their selection
and b) this selection cannot be shown to the user on the instrument
view, the compromise solution was to only take into account and store
the last tube or pixel selected for grouping. In other words, the user
can only sum or extract the last pixel / tube they selected.