Add colour tracking to MuonAnalysis plots
This adds basic color tracking to MuonAnalysis plots so that new lines are plotted using the first unused color in the colormap. This was required so that plots came out with consistent colors when re-plotted and when workspaces were removed.

I debated the best place to put the color_queue and eventually decided it belonged in the view, despite normally trying to keep this as simple as possible as none of the other sections of the code currently care about the color of the lines plotted and it added substantial complexity to put it elsewhere.