Check for null box pointer due to rounding errors
Some scenarios cause getBoxAtCoord to return a nullptr
due to roundoff errors. Protect against this possiblity
when generating a line plot. No test has been added as it
is nontrivial to construct a case and the functionality
fixes a crash in an edge case in MantidPlot sliceviewer
that will be removed in the next release