Tidy up unnecessary search-type flag
This was being used to distinguish between when a search is being run
for autoprocessing vs a manual search initiated by the user. It was only
used when starting autoprocessing to clear the table if a different type
of search had been run. However the two should be interchangeable as
long as other search settings are not changed (which are already checked
for) so I think it is ok to remove this check.

Removing the check means that we do not need to clear the search results
when switching between the two types of search, and this will allow us
to preserve the new annotation columns' info.

Also remove a couple of unused function declarations

Re #29148