Implement SpinBox factory with a Read only line edit
This commit implements a spinbox factory with a read only line edit. This is necessary to avoid a crash on workbench which occurs when a value is entered into the spinbox using the keyboard and `enter` pressed (to submit the value). This occurs as an deleteLater event occurs while the slot from the spinbox is executing, which leads to us accessing a spinbox which has been deleted.

The cause of this delete call is as yet unknown.
Clang format