Set environment for MSVC targets
Currently we generate a visual-studio.bat file
to set the PATH variable to include the third
party directory and then start the IDE. This
avoids having to set user/system-wide environment
variables that could conflict with other things
on developer machines.

These changes set the relevant paths on the targets
such that the generated solutions can then be used
standalone. The visual-studio.bat file has been left
as it is as backup for now. It will likely be removed
in the future.