Always use selected workspace in sliceviewer
Removes functionality that would cause the sliceviewer
to actually use an attached original workspace if
one existed. This made the dynamic rebinning easier
but causes confusion when you bin to non-axis aligned
system and the sliceviewer refuses to show it.

This does remove dynamic rebinning for MDHisto workspaces
and this will be renabled in the future via a different
mechanism. For now the sliceviewer is useless on MD
workspaces that have been binned and sacrificing dynamic
binning is worth it temporarily