Algorithm should now support two methods of input
A vector of strings stores the file names which will either be populated directly or from first and last run values. A new method extracts run numbers from filenames which is then stored in a map which maps filenames to run numbers. This is to support both methods of input. Added a new test for extratRunNumberFromRunName