Created run object
A single dataset in the MuonAnalysis workflow consists of one to many co-added runs. We need to store data for each of these datasets. Previously they were referenced by a list of the constituent runs. This led to problems however as list is not hashable. Here I have converted the object used to referance datasets into a new class. Ideally this should then be used throughout but so as to not bloat the current PR too much I have restricted it's usage to just the MuonGroup and MuonPair classes.

We could also pass the list of runs around as a tuple instead as it does not need to be mutable but on balance I think I prefer the class as it adds to clarity and will let us easily encode the transformation of the run list into a string and back again which is quite a common operation.