Add calculation to invert a symmetric tridiagonal matrix
The error calculation in HistogramData::interpolateYCSplineInplace
requires a matrix to be inverted. This has caused the absorption calculations
in MonteCarloAbsorption to slow down in some cases when spline interpolation
is used (mainly when a sparse instrument calculation is run)
This change implements an analytic calculation to invert a tridiagonal
matrix in the special case where it's symmetric and the diagonal elements
(in each of the 3 diagonals) are the same (if not the inversion function
just reverts to the standard Matrix::Invert function)
The analytic calculation is described here:
"Analytical inversion of symmetric tridiagonal matrices",
G Y Hu and R F O'Connell, J. Phys. A: Math. Gen. 29 (1996)