Fix further problems with tridiagonal matrix inversion
The equality test in the Matrix class didn't work for NaN values so
was resulting in one of my tests passing. I have adjusted Matrix::equals
to sort this out. I have also fixed the problem with the tridiagonal
matrix inversion. The formula in the Hu and O'Connell paper seemed to have
a couple of typos in it for the -2<D<2 case