Fix a crash loading an old batch file in the reflectometry GUI
Batch files from v5.0.0 did not have the background column in the Experiment tab table. Loading one of these files creates a table with the new column count of 10, but only 9 of the columns have a cell created for them. This causes potential crashing e.g. when re-saving the batch, because the Encoder expects 10 valid cells. This commit fixes the problem by adding empty cells if they are not in the file.

Re #29367