Add test for Larmor StateMove throwing if rb is None
Adds a test to check for StateMove throwing when the run number is none.
This happens when we add a user file without a run number in the box.

I checked with the SANS team and it is expected that there x dim is in
m, and y in mm for runs after 2217. So we want to use the same divider
after that to keep the display consistent until they load in run
data.