[SANSILL] bug in validate inputs
During the check of array like parameter, Validateinput() was reporting
a wrong number of parameter (number of comma instead of number of
values). There was also a typo in one of the parameter check.

This commit rewrites the method to check the number of values instead
of the number of commas.