Fix 2d reduction button not working
Fixes the 2D reduction button not working. This was a combination of
minor problems:

- The presenter was not updating the model with the option. This became
a problem when the model was told off for reaching into the view as part
of the TOML work.
- When using a custom user file we ignore the GUI settings. This was a
little too strict as we were ignoring the reduction mode
- For some reason we have the ReductionDimensionality duplicated within
our state objects. If the GUI model does not set both, half the
reduction would be in 2D/1D.