Add function to return wavelength step type as LIN/LOG only
This is slightly messy because the RangeStepType enum is being used directly from the state objects to provide inputs to the workflow algorithm, but the algorithm does not support all of the values in the enum. This should probably be tidied up properly at some point because it is an overly complex way of essentially just specifying whether a -1 should be applied to the rebin params for log steps.

For now, I have fixed the problem by replacing the wavelength_step_type used by the algorithms with a new property wavelength_step_type_property_string, which converts RANGE_LIN/RANGE_LOG to LIN/LOG.

Re #29441