Add ability to enable the MSVC address sanitizer via cmake
Microsoft have recently added an address sanitizer to VS 2019. These
changes allow this to be configured in all of the Mantid targets using
the existing USE_SANITIZER switch that was previously only relevant
to the gcc compiler
Excluded the gmock target from having the sanitizer set up applied because
the gmock cmake set up specifies a W4 logging level that doesn't play
nicely with the sanitizer. Since this isn't our code have assumed we don't
need to sanitize it