Remove Qt.EnableHighDPIScaling attribute
It causes the interfaces not to open on Windows with a dual 4K/2K monitor setup. The
tested scenario was a 4K laptop with a 2K second monitor. Attempting to open any
custom interface produced something that looked open in the task bar but the window
was nowhere to be seen.