Fix remove workspaces in QENS interfaces
This commit fixes the remove button in QENS interfaces. Firstly, the mutable spectra was returned by value and then modified, therefore not updating the stored data. Secondly, the data was removed using a forward iterator, which meant the next index to be removed was invalidated, meaning multiple selection removal would seg fault.
Lastly, the case when all spectra have been removed had to be fixed, in this case we make sure the workspace is removed.