Re #29589 Don't copy det IDs in Qxy
Don't copy detector IDs in QXY, this makes no sense physically and
causes a bug where the first two spectra take the monitor IDs and get
marked as such.

This has been a long standing bug, but was only noticed in Workbench
where monitor spectra are displayed differently (or not at all in colour
fill plots). In Mantid Plot all spectra are plotted identically so this
wasn't noticed