Enable workaround for readling segv in 3.6+
Enables the workaround for Python Segmentation fault in 3.6 and above.
This is the easiest fix, since finding and upstreaming a true bug fix
for PyQt5 would require significant time for (so far) a single module.

----

Technical info:

Readline creates a hook which calls FindModule and passes the result
directly into GetState for the PyState object. However PyQt5 messes up
this internal state by causing the module index tracked internally to be
one less than the modules imported.
This means any calls to FindModule will return a null pointer which
readline happily passes along to GetState.