Add ability to specify multiple attenuation .out files in Pearl config
The Pearl yaml config file in the ISIS powder diffraction scripts now
supports multiple attenuation file paths for .out files via a new
parameter attenuation_files. This parameter is a list of dictionaries
with each dictionary containing a name and path key. One of the files
can be used by specifying a further parameter called attenuation_file
that has a value equal to the name of one of the "attenuation_files"