Add explicit text comment for copy pasting
In the instrument viewer, a text at the bottom tells the user how to
interact with shapes (by clicking, dragging etc...), and now also
instructs him how to copy and paste shapes.

Since this becomes the longest text to be printed there, and since it
does not appear at startup, the minimal size of the window has
accordingly to be enlarged, manually this time, so it is not updated
only when the user switches to another tab. This is done by setting the
minimal width of the instrument display widget itself, since it cannot
be set at the layout level.