Update hard-coded block names in RROLD
ReflectometryReductionOneLiveData fetches live values from the instrument. The block names for the values are currently hard-coded. This commit updates them following a recent change and also supports two versions of the name because instruments are currently switching between the two. Longer term it would be good to avoid hard-coded values.

Re #29269