[DRILL] add and show column through window menu
This commit adds a submenu in the Edit menu to show and hide columns.
This submenu behaves exactly like the context menu. A list of column
title is displayed with a icon representing the visibility state of the
column. By clicking on the name, the visibility of the column is
toggled.