Perform sequential fit by calling fit algortihm
This avoids signals passing between MVP of data and plot after each fit. Now fits are performed and the resuls are passed back in a list to the observer.
Changes also enable sequential fitting button when the fitproperty browser enables the fit button.
Previously an initial fit had to be performed so that the parameters could be accessed via the getFitAlgorithmParameters - now the `live' values are querried directly.
This required exposing several public methods of the fitproperty browser to python. Note that