Add sort order for seq fitting in settings and refactor settings code
Settings code has been refactored to avoid unecessary updating of properties.
There is a change of behaviour when the settings are alterted in the view: previously this would prompt all the settings to be querried and saved to file such that the Cancel button did not actually cancel the changes.
Now the settings are only updated from the view and saved when OK or Applied are clicked.
Have also added labels to clarify the settings for the fitting tab relating to the logs.