Abins: use StringListValidator for parameters
This gets a nice drop-down list of options. However, not all the
options will be valid for the current instrument. I am not aware of a
way to limit this with a simple Python Algorithm GUI; but we may be
able to provide some validation information so that users no if an
incompatible option is selected.