Add API function to replace a viewed workspace
API function in the presenter (though the MVP model is not perfect here)
to allow the user to replace the workspace that is currently being shown
in an instrument view.

BUT it only works if the instrument view is called within the thread
were it was created (in other words, use it with caution on threading
situations).