Prevent EqualBinsChecker generating unwanted log messages
The EqualBinsChecker class was generating error\warning messages during the
lazy algorithm initialisation where doc strings are generated for each algorithm
(if there was a workspace present in the ADS with uneven bins). The messages
were being triggered from the MaxEnt and ApplyDeadTimeCorr algorithms
I've changed this so when EqualBinsChecker is called from the IValidator
it only returns a true\false statement about whether it's valid or not
The extra behaviour where EqualBinsChecker performs a softer warning check
can now only be trigged by explicitly creating an EqualBinsChecker object
in an algorithm (as was already done in the FFT algorithm). I have updated the
MaxEnt algorithm so it applies this additional check. ApplyDeadTimeCorr only
checked for errors so didn't require this extra bit of code