Remove algorithm cache in docs directive
The cache causes incorrect values to be placed
into the properties tables as previous initialized
algorithms are used that have already had properties
set on them but not cleared.
The cache does not warrant enough of a speed up over
creating/initializing a new algorithm each time to
attempt to clear the existing property values.