Replace unique_ptr with QPointer
Pressing Esc on a QDialog closes the dialog
but crucially closeEvent is not called.
The delete on close attribute causes the
dialog to be deleted on pressing Esc but
this would leave the unique_ptr with a dangling
pointer. QPointer nullifies itself when it's
object is deleted and manages this lifecycle
for us