Fix problem with PaalmanPingsMonteCarloAbsorptionTest on Windows
The PaalmanPingsMonteCarloAbsorptionTest.FlatPlateTest was failing
intermittently on Windows. The cause of this was two things:

a) there was a problem in the duplicates check when adding points into a
track in Track::addPoint which meant that sometimes a point wasn't added
even though no duplicate existed already. I think this problem has
perhaps come about because the track uses a boost smallvector which
seems to have some slots pre-initialised
b) there is a vector of V3D points that is populated in
CSGObject::interceptSurface. The sort order of this wasn't well defined
meaning that a) produced different results on subsequent runs of the
same system test. The sort order is driven by the order of some pointers
to Surfaces in the CSGObject::m_surList member
I have fixed the bug in a) and have removed the sort on CSGObject::m_surList

The fix has changed the system test reference file slightly