Refactor XIntegration use in the instrument widget
The Xintegration widget, which control the integration scroll at the
bottom of the instrument view, should only appear when the workspace
needs it.

This was currently the case, but only if the instrument view was opened
with this workspace : changing the underlying workspace midway, as
intended by replaceWorkspace, was not managed.

Thus, the way the integration range is managed was changed to a more
concise and modular way, so that it is only hidden or shown when needed,
but exists in all cases.